Add plume agent init script ( RDKB-4761 )
Add an init script for the plume agent that runs the plume manager if
SONAdminStatus and SONOperationalStatus are both true.

avoid running the opensync agent before WiFi is initialized ( RDKB-5029 )

If the opensync agent is run before WiFi is initialized, the gateway
will stay offline on the Plume NOC.

plume_init.sh: retry failed RPC commands ( OFW-2009 )

Fix issue causing opensync not to start during bootup

Sometimes during bootup, when the ARM CPU is busy the rpcclient
commands used to get the opensync syscfg variables fail, causing
opensync to not be started.

Fix: Retry the rpcclient commands when they fail.