fix setting initial OVS mode to true when SONAdminStatus is enabled ( OFW-4716 F3896SI-1506 F3896SI-2041 )
[Issue]
CPE is offline in Plume NOC, when moved from IPv4 to DUAL and Dual to IPv4 mode using config file

[Solution]
When we upgrade the firmware using bootconfig file son_operational_status and son_admin_status syscfg variables are true,
but the mesh variables don't exist and the mesh agent will not triggers the managers.init script which actually starts the opensync process.
As a fix we are checking the son_operational_status and son_admin_status and if it is set to true we are setting initial OVS mode to true which will further start the plume manager.

[TESTING]

Steps:
1. Provision DUT in IPv4 mode using plume enabled config file
2. Verified the status of CPE is online in Plume NOC
3. Move DUT to DUAL mode using plume enabled config file
4. Verified the status of CPE is online in Plume NOC
5. Move DUT to IPv4 mode using plume enabled config file
6. Verified the status of CPE is online in Plume NOC

Upstream Candidate: Yes

Signed-off-by: arudas <arudas.ext@libertyglobal.com>