[PATCH PORTING]: After changing WiFi setting, POD can no longer connect to the Gateway (OFW-4603).
[Issue]
Backhaul is being disabled

[Description]
As soon the box reboots, mesh-agent disables the backhauls, which is pointed by the above comment.
Opensync is responsible to enable back.disabling is done without the applysetting. configs are held in HAL.
This is why still opensync see backhauls as enabled.
When there is a change from the UI, Applysetting is called, the older config held in HAL is Applied. this makes backhaul to disable.

[Solution]
Do not check status of backhaul after reboot, disable directly. Add an Applysetting after disabling.