plume_init.sh: set the 5GHz channel number 44 after device reboot if plume is enabled ( OFW-5163 )
[Issue]
Plume optimization lasts longer than expected.
CPE always selects DFS channel and plume overwrites to 44 as part of optimization.

[Description]
To implement the ECO-2838
This is done to cannel avoid optimization from plume to select 44 after reboot.
With this change one optmization from plume is avoided after device reboot.

This is as part of ECO-2838(Only part 1 is implemented, part is done later, when the solution is added as a feature for all platforms).
part 1: Select channel 44 on device reboot when plume is enabled and reset when plume is disabled.
part 2: Select the 2.4GHz channel that was selected by plume earlier. TODO

[Solution]
- Set the 5GHz channel number 44 after device reboot, when plume is enabled.
- Unset channel 44 channel, when plume is disabled.

[Tesing performed]
Verify device selects channel 44 on reboot. Also when plume enabled after boot.
Verify device enabled ACS and unsets channel 44 after plume is disabled.

Signed-off-by: Sidda(sihiregoudar@libertyglobal.com)